Revert "[DEX-20402] - Platform link and remove dep from scripts.js"
