+++ _src/blocks/exit-popup/exit-popup.css
@@ -8,9 +8,10 @@ main .exit-popup-container {
   transform: translate(-50%, -50%);
   border-radius: 10px;
   background-color: white;
+  padding: 1em;
   color: black;
   padding: 0;
-  box-shadow: 0 3px 6px #00000029;
+  box-shadow: 0px 3px 6px #00000029;
   display: none;
 }
