+++ _src/blocks/fragment-metadata/fragment-metadata.js
@@ -1,5 +1,5 @@
 import { readBlockConfig } from '../../scripts/lib-franklin.js';
-import { openUrlForOs } from '../../scripts/utils/utils.js';
+import { openUrlForOs } from '../../scripts/scripts.js';
 
 export default function decorate(block) {
   const {
