--- _src/blocks/platform-link/platform-link.js
@@ -1,15 +1,0 @@
-import { getDatasetFromSection, openUrlForOs } from '../../scripts/utils/utils.js';
-
-export default function decorate(block) {
-  block.classList.add('default-content-wrapper');
-  const blockDataset = getDatasetFromSection(block);
-
-  const {
-    linkandroid, linkios, linkmacos, linkwindows,
-  } = blockDataset;
-
-  const freeDownloadButton = block.querySelector('a');
-  if (freeDownloadButton) {
-    openUrlForOs(linkmacos, linkwindows, linkandroid, linkios, freeDownloadButton);
-  }
-}
